requirements.txt
aioredis==1.1.0
asgi-redis==1.4.3
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
certifi==2018.4.16
channels==2.1.2
channels-redis==2.2.1
constantly==15.1.0
daphne==2.2.1
Django==2.0.7
django-crontab==0.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
pipenv==2018.7.1
PyHamcrest==1.9.0
PyJWT==1.6.4
pytz==2018.5
redis==2.10.6
six==1.11.0
Twisted==18.7.0
txaio==2.10.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
zope.interface==4.5.0
